requirements.txt
flask >= 0.12
flask_restplus >= 0.10.1
requests >= 2.18
Flask-Testing >= 0.7
wechatpy >=1.7.0
cryptography>=0.8.2
mysqlclient >= 1.3.8
mysql-connector >= 2.1
peewee >= 3.5.0